removed ef package, cause it used in StandartCRUD lib + api template + sonar (Profiles.BLL/Interfaces renamed folder, it was Interrfaces => changed usings)
