+++ Clinic.API.sln
@@ -27,6 +27,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Profiles.DAL", "src\Profile
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StandartCRUD", "common\StandartCRUD\StandartCRUD.csproj", "{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services.API", "src\Services\Services.API\Services.API.csproj", "{70CD9825-897A-48C5-88D0-76C2BABEC50D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services.BLL", "src\Services\Services.BLL\Services.BLL.csproj", "{CB570953-A11E-4589-9AC1-7EED58735057}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services.DAL", "src\Services\Services.DAL\Services.DAL.csproj", "{EFCA602C-EC27-474F-A731-3481E80DB010}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -83,6 +89,18 @@ Global
 		{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70CD9825-897A-48C5-88D0-76C2BABEC50D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70CD9825-897A-48C5-88D0-76C2BABEC50D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70CD9825-897A-48C5-88D0-76C2BABEC50D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70CD9825-897A-48C5-88D0-76C2BABEC50D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CB570953-A11E-4589-9AC1-7EED58735057}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CB570953-A11E-4589-9AC1-7EED58735057}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB570953-A11E-4589-9AC1-7EED58735057}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CB570953-A11E-4589-9AC1-7EED58735057}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EFCA602C-EC27-474F-A731-3481E80DB010}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EFCA602C-EC27-474F-A731-3481E80DB010}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EFCA602C-EC27-474F-A731-3481E80DB010}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EFCA602C-EC27-474F-A731-3481E80DB010}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
