+++ src/Profiles/Profiles.BLL/Interfaces/IAccountService.cs
@@ -1,7 +1,7 @@
 ﻿using Profiles.BLL.Models;
 using StandartCRUD.StandartBLL;
 
-namespace Profiles.BLL.Interrfaces
+namespace Profiles.BLL.Interfaces
 {
     public interface IAccountService : IGenericService<Account>
     {
