+++ src/Profiles/Profiles.BLL/Services/DoctorService.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using Profiles.BLL.Interrfaces;
+using Profiles.BLL.Interfaces;
 using Profiles.BLL.Models;
 using Profiles.DAL.Entities;
 using Profiles.DAL.Interfaces;
@@ -7,8 +7,8 @@ using StandartCRUD.StandartBLL;
 
 namespace Profiles.BLL.Services
 {
-    public class DoctorService(IDoctorRepository doctorRepository, IMapper mapper) : 
-        GenericService<DoctorEntity, Doctor>(doctorRepository, mapper), 
+    public class DoctorService(IDoctorRepository doctorRepository, IMapper mapper) :
+        GenericService<DoctorEntity, Doctor>(doctorRepository, mapper),
         IDoctorService
     {
         private readonly IDoctorRepository _doctorRepository = doctorRepository;
