+++ src/Profiles/Profiles.BLL/Services/ReceptionistService.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using Profiles.BLL.Interrfaces;
+using Profiles.BLL.Interfaces;
 using Profiles.BLL.Models;
 using Profiles.DAL.Entities;
 using Profiles.DAL.Interfaces;
@@ -7,8 +7,8 @@ using StandartCRUD.StandartBLL;
 
 namespace Profiles.BLL.Services
 {
-    public class ReceptionistService(IReceptionistRepository receptionistRepository, IMapper mapper) : 
-        GenericService<ReceptionistEntity, Receptionist>(receptionistRepository, mapper), 
+    public class ReceptionistService(IReceptionistRepository receptionistRepository, IMapper mapper) :
+        GenericService<ReceptionistEntity, Receptionist>(receptionistRepository, mapper),
         IReceptionistService
     {
         private readonly IReceptionistRepository _receptionistRepository = receptionistRepository;
