+++ src/Profiles/Profiles.BLL/Services/SpecializationService.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using Profiles.BLL.Interrfaces;
+using Profiles.BLL.Interfaces;
 using Profiles.BLL.Models;
 using Profiles.DAL.Entities;
 using Profiles.DAL.Interfaces;
@@ -7,8 +7,8 @@ using StandartCRUD.StandartBLL;
 
 namespace Profiles.BLL.Services
 {
-    public class SpecializationService(ISpecializationRepository specializationRepository, IMapper mapper) : 
-        GenericService<SpecializationEntity, Specialization>(specializationRepository, mapper), 
+    public class SpecializationService(ISpecializationRepository specializationRepository, IMapper mapper) :
+        GenericService<SpecializationEntity, Specialization>(specializationRepository, mapper),
         ISpecializationService
     {
         private readonly ISpecializationRepository _specializationRepository = specializationRepository;
