+++ src/Services/Services.API/Services.API.http
@@ -1,0 +1,6 @@
+@Services.API_HostAddress = http://localhost:5159
+
+GET {{Services.API_HostAddress}}/
+Accept: application/json
+
+###
